refactor(editor): キーボード追従を lib に出す
「画面の下に貼り付くものをキーボードの上へ逃がす」のは記法ツールバー
だけの話ではない。テンプレ編集の変数チップも同じ場所に置く以上、同じ
逃がし方が要る。

MarkdownToolbar から取ると Milkdown と ProseMirror が一緒に付いてくる。
入り口を lib に移して、キーボードの都合だけを import できるようにする。

挙動は変えていない。判定のテストはそのまま移した。
