+++ tauri-app/src/components/MarkdownToolbar.test.tsx
@@ -2,7 +2,7 @@ import { describe, it, expect, vi, afterEach } from "vitest";
 import { render, screen, fireEvent, cleanup } from "@solidjs/testing-library";
 import { createSignal } from "solid-js";
 import type { Editor } from "@milkdown/kit/core";
-import MarkdownToolbar, { keyboardTop } from "./MarkdownToolbar";
+import MarkdownToolbar from "./MarkdownToolbar";
 
 function createMockEditor(): Editor {
   return {
@@ -67,27 +67,5 @@ describe("MarkdownToolbar", () => {
 
     setEditor(undefined);
     expect(screen.queryByRole("toolbar")).toBeNull();
-  });
-});
-
-describe("keyboardTop", () => {
-  const WINDOW_HEIGHT = 900;
-
-  // Android では閉じていても visualViewport がナビゲーションバーぶんを含んだ
-  // 全高を返すので、その値で固定するとツールバーがバーの裏に潜り込む
-  it("returns nothing while the keyboard is closed", () => {
-    expect(keyboardTop({ offsetTop: 0, height: WINDOW_HEIGHT }, WINDOW_HEIGHT)).toBeUndefined();
-  });
-
-  it("ignores a shrink too small to be a keyboard", () => {
-    expect(keyboardTop({ offsetTop: 0, height: 860 }, WINDOW_HEIGHT)).toBeUndefined();
-  });
-
-  it("returns the keyboard top once it opens", () => {
-    expect(keyboardTop({ offsetTop: 0, height: 500 }, WINDOW_HEIGHT)).toBe(500);
-  });
-
-  it("follows the viewport while it is scrolled", () => {
-    expect(keyboardTop({ offsetTop: 120, height: 500 }, WINDOW_HEIGHT)).toBe(620);
   });
 });
