+++ tauri-app/src/components/MarkdownToolbar.tsx
@@ -1,4 +1,4 @@
-import { Show, createSignal, onMount, onCleanup } from "solid-js";
+import { Show, onMount, onCleanup } from "solid-js";
 import { Portal } from "solid-js/web";
 import type { Editor } from "@milkdown/kit/core";
 import { commandsCtx, editorViewCtx, rootCtx } from "@milkdown/kit/core";
@@ -11,6 +11,7 @@ import {
 import type { Command } from "@milkdown/kit/prose/state";
 import { deleteCurrentBlock, exitCodeBlock } from "../lib/block-commands";
 import { t } from "../lib/i18n";
+import { createKeyboardTop, keyboardTopStyle } from "../lib/keyboard";
 import Icon from "./Icon";
 import type { JSX } from "solid-js";
 
@@ -18,28 +19,8 @@ interface MarkdownToolbarProps {
   editor: Editor | undefined;
 }
 
-/** これ以下の縮みはスクロールバーや URL バーの誤差で、キーボードとは見なさない。 */
-const KEYBOARD_MIN_HEIGHT = 100;
-
-/**
- * キーボードの上端。閉じているあいだは `undefined` を返し、CSS の
- * `bottom: var(--safe-bottom)` に任せる。
- *
- * Android では閉じていても `visualViewport.height` がナビゲーションバーを含んだ
- * 全高になるため、その値で `top` を固定するとツールバーがバーの裏に潜り込む。
- */
-export function keyboardTop(
-  viewport: { offsetTop: number; height: number },
-  windowHeight: number,
-): number | undefined {
-  if (viewport.height >= windowHeight - KEYBOARD_MIN_HEIGHT) {
-    return undefined;
-  }
-  return viewport.offsetTop + viewport.height;
-}
-
 export default function MarkdownToolbar(props: MarkdownToolbarProps): JSX.Element {
-  const [toolbarTop, setToolbarTop] = createSignal<number | undefined>();
+  const toolbarTop = createKeyboardTop();
 
   // ツールバーが出ている間(=編集中)は下部タブを隠す。fixed のツールバーが
   // タブに重なって Timeline / Notes が押せない・誤タップでモードが変わる、の
@@ -47,24 +28,6 @@ export default function MarkdownToolbar(props: MarkdownToolbarProps): JSX.Elemen
   onMount(() => {
     document.body.classList.add("md-toolbar-open");
     onCleanup(() => document.body.classList.remove("md-toolbar-open"));
-  });
-
-  onMount(() => {
-    const vv = window.visualViewport;
-    if (!vv) {
-      return;
-    }
-
-    const update = () => {
-      setToolbarTop(keyboardTop(vv, window.innerHeight));
-    };
-
-    vv.addEventListener("resize", update);
-    vv.addEventListener("scroll", update);
-    onCleanup(() => {
-      vv.removeEventListener("resize", update);
-      vv.removeEventListener("scroll", update);
-    });
   });
 
   const exec = (run: (editor: Editor) => void) => {
@@ -90,8 +53,6 @@ export default function MarkdownToolbar(props: MarkdownToolbarProps): JSX.Elemen
     );
   };
 
-  const top = () => toolbarTop();
-
   return (
     <Show when={props.editor}>
       <Portal>
@@ -99,11 +60,7 @@ export default function MarkdownToolbar(props: MarkdownToolbarProps): JSX.Elemen
           class="markdown-toolbar"
           role="toolbar"
           aria-label="Markdown formatting"
-          style={
-            top() === undefined
-              ? undefined
-              : { top: `${top()}px`, bottom: "auto", transform: "translateY(-100%)" }
-          }
+          style={keyboardTopStyle(toolbarTop())}
         >
           <button
             type="button"
